@@ -27,12 +27,12 @@
         }
     },
     "require": {
-        "php": "^7.1",
-        "illuminate/database": "5.7.* || 5.8.*",
-        "illuminate/http": "5.7.* || 5.8.*",
-        "illuminate/routing": "5.7.* || 5.8.*",
-        "illuminate/support": "5.7.* || 5.8.*",
-        "illuminate/contracts": "5.7.* || 5.8.*",
-        "guzzlehttp/guzzle": "^6.3"
+        "php": "^7.4|^8.0",
+        "illuminate/database": "^8.0.0",
+        "illuminate/http": "^8.0.0",
+        "illuminate/routing": "^8.0.0",
+        "illuminate/support": "^8.0.0",
+        "illuminate/contracts": "^8.0.0",
+        "guzzlehttp/guzzle": "^7.3.0"
     }
 }
