@@ -2,48 +2,48 @@
 
 namespace Alish\Telegram\API;
 
-class CallbackQuery extends BaseTelegram
+class CallbackQuery extends TelegramAPI
 {
 
     /**
      * @var string $id
      * Unique identifier for this query
      */
-    protected $id;
+    public string $id;
 
     /**
      * @var User $from
      * Telegram
      */
-    protected $from;
+    public User $from;
 
     /**
      * @var Message|null $message
      * Optional. Message with the callback button that originated the query. Note that message content and message date will not be available if the message is too old
      */
-    protected $message;
+    public ?Message $message;
 
     /**
      * @var string|null $inline_message_id
      * Optional. Identifier of the message sent via the bot in inline mode, that originated the query.
      */
-    protected $inline_message_id;
+    public ?string $inline_message_id;
 
     /**
      * @var string $chat_instance
      * Global identifier, uniquely corresponding to the chat to which the message with the callback button was sent. Useful for high scores in games.
      */
-    protected $chat_instance;
+    public string $chat_instance;
 
     /**
-     * @var string $data
+     * @var string|null $data
      * Optional. Data associated with the callback button. Be aware that a bad client can send arbitrary data in this field.
      */
-    protected $data;
+    public ?string $data;
 
     /**
-     * @var string $game_short_name
+     * @var string|null $game_short_name
      * Optional. Short name of a Game to be returned, serves as the unique identifier for the game
      */
-    protected $game_short_name;
+    public ?string $game_short_name;
 }
