@@ -2,12 +2,12 @@
 
 namespace Alish\Telegram\API;
 
-class InlineKeyboardMarkup extends BaseTelegram
+class InlineKeyboardMarkup extends TelegramAPI
 {
 
     /**
      * @var InlineKeyboardButton[][] $inline_keyboard
      * Array of button rows, each represented by an Array of InlineKeyboardButton objects
      */
-    protected $inline_keyboard;
+    protected array $inline_keyboard;
 }
