@@ -2,7 +2,7 @@
 
 namespace Alish\Telegram\API;
 
-class PhotoSize extends BaseTelegram
+class PhotoSize extends TelegramAPI
 {
 
 
@@ -10,23 +10,29 @@ class PhotoSize extends BaseTelegram
      * @var string $file_id
      * Unique identifier for this file
      */
-    protected $file_id;
+    public string $file_id;
+
+    /**
+     * @var string $file_unique_id
+     * Unique identifier for this file, which is supposed to be the same over time and for different bots. Can't be used to download or reuse the file
+     */
+    public string $file_unique_id;
 
     /**
      * @var integer $width
      * Photo width
      */
-    protected $width;
+    public int $width;
 
     /**
      * @var integer $height
      * 	Photo height
      */
-    protected $height;
+    public int $height;
 
     /**
      * @var integer|null $file_size
      * Optional. File size
      */
-    protected $file_size;
+    public ?int $file_size;
 }
