@@ -2,14 +2,14 @@
 
 namespace Alish\Telegram\API;
 
-class ReplyKeyboardRemove extends BaseTelegram
+class ReplyKeyboardRemove extends TelegramAPI
 {
     /**
      * @var true $remove_keyboard
      * Requests clients to remove the custom keyboard (user will not be able to summon this keyboard;
      * if you want to hide the keyboard from sight but keep it accessible, use one_time_keyboard in ReplyKeyboardMarkup)
      */
-    protected $remove_keyboard;
+    public $remove_keyboard;
 
     /**
      * @var boolean|null $selective
@@ -19,5 +19,5 @@ class ReplyKeyboardRemove extends BaseTelegram
      * Example: A user votes in a poll, bot returns confirmation message in reply to the vote and removes the keyboard for that user,
      * while still showing the keyboard with poll options to users who haven't voted yet.
      */
-    protected $selective;
+    public ?bool $selective;
 }
