@@ -2,18 +2,18 @@
 
 namespace Alish\Telegram\API;
 
-class UserProfilePhotos extends BaseTelegram
+class UserProfilePhotos extends TelegramAPI
 {
 
     /**
      * @var integer $total_count
      * Total number of profile pictures the target user has
      */
-    protected $total_count;
+    protected int $total_count;
 
     /**
      * @var PhotoSize[][] $photos
      * Requested profile pictures (in up to 4 sizes each)
      */
-    protected $photos;
+    protected array $photos;
 }
