@@ -8,7 +8,7 @@
 
 namespace Alish\Telegram\API;
 
-class MaskPosition extends BaseTelegram
+class MaskPosition extends TelegramAPI
 {
     /**
      * @var string
